refactor: move favorite teachers module to submodule
