+++ src/modules/student/submodules/favorite-teachers/domain/dtos/favorite-teacher.dto.ts
