+++ src/modules/student/submodules/favorite-teachers/domain/entities/favorite-teachers.entity.ts
