+++ src/modules/student/submodules/favorite-teachers/infra/repositories/favorite-teachers.repository.ts
